+++ packages/node-utils/index.ts
@@ -1,1 +1,2 @@
 export * from './dirname';
+export * from './filename';